refactor: view 의존성 여부에 따라 model과 controller 로직 분리 및 폴더 이동
