+++ __tests__/entities/Car.test.js
@@ -1,5 +1,5 @@
-import Car from '../../src/Models/Car/Car.js';
-import MoveStrategy from '../../src/Models/MoveStrategy/MoveStrategy.js';
+import Car from '../../src/Models/entities/Car/Car.js';
+import MoveStrategy from '../../src/Models/entities/MoveStrategy/MoveStrategy.js';
 
 describe('Car 생성자 테스트', () => {
     describe('생성자 내부 로직 테스트', () => {
