+++ src/Models/entities/MoveStrategy/errors.js
