+++ src/Models/service/ValidationError.js
@@ -1,8 +1,8 @@
-import RuntimeError from '../RuntimeError.js';
+import RuntimeError from '../../RuntimeError.js';
 
 // 사용자 입력값 검증 오류
 export default class ValidationError extends RuntimeError {
-    static #type = 'USER_INPUT';
+    static #type = 'SERVICE';
 
     constructor(message) {
         super(ValidationError.#type, message);
